LIBITD-2254. Added baseline Alembic migration
Added a baseline Alembic migration that creates the necessary tables,
indices, and views for an empty database.

This enables Alembic to replace the PATSy "schema" command, and provides
the base for future database migrations.

The migration has been set up to work, even if the tables already exist.

A naming convention for database primary and foreign keys was added
to "patsy/model.py", and any existing primary and foreign keys are
renamed (via dropping and recreating them) as part of the migration.

https://umd-dit.atlassian.net/browse/LIBITD-2254