LIBITD-2254. Removed PATSy "schema" command
Removed the PATSy "schema" command, as it is no longer needed, and
could interfere with the Alembic database migration tool.

Updated the documentation to remove the "schema" command, and add
information about Alembic.

In the tests, replaced the use of the "schema" command to update the
database with "init_database" and "clear_database" helper functions
in the "test/__init__.py" file.

https://umd-dit.atlassian.net/browse/LIBITD-2254